+++ lib/straw.rb
@@ -5,6 +5,14 @@ require_relative "straw/version"
 module Straw
   class Error < StandardError; end
 
+  def self.logger
+    @logger ||= Logger.new($stderr, level: ENV.fetch("LOG_LEVEL", Logger::INFO)).tap do |x|
+      x.formatter = proc do |_severity, _datetime, _progname, message|
+        "[#{VERSION}] #{message}\n"
+      end
+    end
+  end
+
   module Memoizable
     def memoize(key)
       if memoized?(key)