+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - feat: Add Memoizable module
+- feat: Add default logger
 
 ## [0.1.0] - 2021-12-22